+++ src/cli/commands/analyze/dependency-risk-helpers/risk-filter.ts
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import type { EffectiveStatus, RiskVM } from './view-model/dependency-risks-view-model.ts';
+import type { EffectiveStatus, RiskVM } from './view-model';
 
 export type RiskFilterPredicate = (risk: RiskVM) => boolean;
