+++ src/cli/commands/analyze/dependency-risk-helpers/table/format-table-vulnerability-group.ts
@@ -18,10 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import type {
-  VulnerabilityGroupVM,
-  VulnerabilityRiskVM,
-} from '../view-model/dependency-risks-view-model.ts';
+import type { VulnerabilityGroupVM, VulnerabilityRiskVM } from '../view-model';
 import { LINE_INDENT } from './format-dependency-risks-table.ts';
 import { formatRecommendationLine } from './format-table-recommendation.ts';
 import { genericRiskInfo } from './format-table-risk-generic.ts';
