+++ src/cli/commands/analyze/dependency-risk-helpers/table/index.ts
@@ -18,14 +18,4 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import type { MalwareGroupVM } from '../view-model';
-import { LINE_INDENT } from './format-dependency-risks-table.ts';
-import { formatRecommendationLine } from './format-table-recommendation.ts';
-import { genericRiskInfo } from './format-table-risk-generic.ts';
-
-export function appendMalwareGroup(lines: string[], group: MalwareGroupVM): void {
-  for (const risk of group.selectedRisks) {
-    lines.push(`${LINE_INDENT}${genericRiskInfo(risk, 'Malicious package')}`);
-  }
-  lines.push(`${LINE_INDENT}${formatRecommendationLine(group.recommendation)}`);
-}
+export * from './format-dependency-risks-table.ts';
